🐛 macros: ignore r# prefix in method names
The r# prefix required for identifiers that would otherwise be
rust keywords is no longer carried over into the dbus
method / property / signal names when used in
interface or proxy methods.

Users no longer have to resort to the "name" attribute in these cases.

There is a related open issue regarding r# in parameter names (#158).
That is not addressed here, but might be in a follow-up PR.

Closes #214